requirements.txt
PyQt5>=5.15.0
opencv-python>=4.8.0
numpy>=1.24.0
python-vlc>=3.0.18

# Subtítulos automáticos con Whisper
openai-whisper
torch  # CPU — para GPU Nvidia: pip install torch --index-url https://download.pytorch.org/whl/cu118
